Factor out code that can be shared by CCE and CCE+GH
Also use cce_step_choosers in CharacteristicEvolution component.
